+++ config.js
@@ -1,5 +1,5 @@
 var config = {     // REMEMBER TO CHANGE is_production and is_simulator!!!!!
-    version				:  '1.9.13', // Zendesk Support within app
+    version				:  '1.9.14', // Rho start/end values in playlist editing, WC bugfixes
     debug         : false,
     env           : process.env.NODE_ENV,
     matt: {